+++ api/configuration/configuration.go
@@ -8,7 +8,7 @@ type SecretKeys struct {
 
 // Configuration interface
 type Configuration struct {
-	BackendAddr  string `toml:"backend_addr,omitempty" mapstructure:"backend_addr,omitempty"`
+	BackendAddress  string `toml:"backend_address,omitempty" mapstructure:"backend_addr,omitempty"`
 	Keys         SecretKeys
 	AccessToken  string `toml:"access_token,omitempty" mapstructure:"access_token,omitempty"`
 	EnvName      string `toml:"envName,omitempty" mapstructure:"envName,omitempty"`
