+++ internal/service/client.go
@@ -22,7 +22,7 @@ func grpcClient(ctx *context.Context) (*grpc.ClientConn, *context.Context, error
 		opts = append(opts, grpc.WithTransportCredentials(cred))
 	}
 
-	conn, err := grpc.Dial(appConfig.BackendAddr, opts...)
+	conn, err := grpc.Dial(appConfig.BackendAddress, opts...)
 
 	if err != nil {
 		return nil, nil, err
