+++ internal/service/configure.go
@@ -6,10 +6,10 @@ import (
 	auth "github.com/dream11/odin/proto/gen/go/dream11/od/auth/v1"
 )
 
-// Environment performs operation on environment like create, list, describe, delete
+// Configure used to perform odin congigure
 type Configure struct{}
 
-// ListEnvironments List environments
+// GetUserToken Get User Token
 func (c *Configure) GetUserToken(ctx *context.Context, request *auth.GetUserTokenRequest) (*auth.GetUserTokenResponse, error) {
 	conn, requestCtx, err := grpcClient(ctx)
 	if err != nil {
